Move ValueRef::from_value
